+++ src/hooks/block-subscription/use-block-subscription.hook.ts
@@ -1,21 +1,11 @@
-import { BlockResponse, BlockFullHeader } from '@taquito/rpc';
 import { useEffect, useRef, useState } from 'react';
 
-export interface BlockInterface extends Pick<BlockResponse, 'protocol' | 'chain_id' | 'hash'> {
-  header: Pick<BlockFullHeader, 'level' | 'timestamp'>;
-}
+import { TEZOS_DEXES_API_URL } from '../../screens/swap/config';
+import { BlockInterface, EMPTY_BLOCK } from './block.interface';
 
-export const EMPTY_BLOCK: BlockInterface = {
-  protocol: '',
-  chain_id: '',
-  hash: '',
-  header: {
-    level: 0,
-    timestamp: ''
-  }
-};
+const BLOCK_UPDATE_API_URL = `${TEZOS_DEXES_API_URL}/block`;
 
-export const useBlockSubscription = (webSocketUrl: string) => {
+export const useBlockSubscription = () => {
   const webSocketRef = useRef<WebSocket>();
   const refreshControlRef = useRef(Math.random());
 
@@ -29,10 +19,9 @@ export const useBlockSubscription = (webSocketUrl: string) => {
   };
 
   useEffect(() => {
-    webSocketRef.current = new WebSocket(webSocketUrl);
+    webSocketRef.current = new WebSocket(BLOCK_UPDATE_API_URL);
 
-    webSocketRef.current.onerror = (errorEvent: Event) => {
-      console.log(errorEvent);
+    webSocketRef.current.onerror = () => {
       setHasFailed(true);
       setIsRefreshing(false);
     };
@@ -46,7 +35,7 @@ export const useBlockSubscription = (webSocketUrl: string) => {
     };
 
     return () => webSocketRef.current?.close();
-  }, [webSocketUrl, refreshControlRef.current]);
+  }, [refreshControlRef.current]);
 
   return { block, hasFailed, isRefreshing, onRefresh };
 };
